zstd - refactoring

+++ src/main/java/ru/olegcherednik/zip4jvm/io/zstd/FrameHeader.java
@@ -18,13 +18,13 @@ import java.util.StringJoiner;
 
 class FrameHeader
 {
-    final long headerSize;
+    final int headerSize;
     final int windowSize;
     final long contentSize;
     final long dictionaryId;
     final boolean hasChecksum;
 
-    public FrameHeader(long headerSize, int windowSize, long contentSize, long dictionaryId, boolean hasChecksum)
+    public FrameHeader(int headerSize, int windowSize, long contentSize, long dictionaryId, boolean hasChecksum)
     {
         this.headerSize = headerSize;
         this.windowSize = windowSize;
